🐞[fix] profile ui 수정 및 profile, member entity 수정

See merge request S2uJeong1/bucketlist!74

+++ bucket_list/src/main/java/com/team9/bucket_list/domain/entity/MemberReview.java
@@ -6,8 +6,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.io.Serializable;
-
 @Entity
 @Builder
 @NoArgsConstructor
@@ -24,7 +22,7 @@ public class MemberReview {
     @Column(length = 500)
     private String content;
     private Long writerId;
-    private int rate;
+    private double rate;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
